requirements.txt
accelerate==1.3.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
Automat==24.8.1
bcrypt==4.2.1
blinker==1.9.0
blis==1.2.0
cachetools==5.5.1
captcha==0.6.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
cryptography==44.0.1
cssselect==1.2.0
cymem==2.0.11
defusedxml==0.7.1
distro==1.9.0
docx==0.2.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
extra-streamlit-components==0.1.71
faiss-cpu==1.10.0
filelock==3.17.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
# intel-openmp==2021.4.0
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
# mkl==2021.4.0
mpmath==1.3.0
murmurhash==1.0.12
narwhals==1.26.0
networkx==3.4.2
nltk==3.9.1
notion-client==2.2.1
numpy==1.26.4
openai==1.63.2
packaging==24.2
pandas==2.2.3
parsel==1.10.0
pillow==11.1.0
preshed==3.0.9
Protego==0.4.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
PyYAML==6.0.2
queuelib==1.7.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.12.0
sentence-transformers==3.4.1
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.42.0
streamlit-authenticator==0.4.1
streamlit-option-menu==0.4.0
sympy==1.13.1
# tbb==2021.13.1
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.5.0
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
torch==2.3.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.3
Twisted==24.11.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
w3lib==2.3.1
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
wrapt==1.17.2
zope.interface==7.2
